+++ datasentry-management/src/main/java/com/touhouqing/datasentry/cleaning/model/CleaningJob.java
@@ -22,7 +22,7 @@ public class CleaningJob {
 
 	private Long agentId;
 
-	private Integer datasourceId;
+	private Long datasourceId;
 
 	private String tableName;
